tmp/challenge/medium_negotiation/requirements.txt
Pillow==9.0.1
flask
Flask-Limiter
gunicorn
markdown
filetype
